P0.0: Extract DirectoryTraversal.cs
Pure mechanical move, no behavior change: ExcludedDirectoryNames,
DirectoryWalkOptions, IsAlwaysExcludedFile, EnumerateFiles, MatchesAny, and
CompilePatterns move out of ScanEngine.cs into a new DirectoryTraversal
class, mirroring the same extraction being done in the sibling
LineEndingNormalizer project as part of a shared hardening pass across both.

ScanEngine.ScanDirectory now calls DirectoryTraversal.CompilePatterns/
EnumerateFiles instead of its own private methods; no other behavior change.
